requirements.txt
Pillow
exifread
python-docx
PyPDF2
rich
streamlit

